# expression: implement vectorized evaluation for `builtinTimestampAddSig`