--- a/expression/builtin_time_vec_test.go
+++ b/expression/builtin_time_vec_test.go
@@ -84,7 +84,7 @@ var vecBuiltinTimeCases = map[string][]vecExprBenchCase{
 		{
 			retEvalType:   types.ETString,
 			childrenTypes: []types.EvalType{types.ETString, types.ETInt, types.ETDatetime},
-			geners:        []dataGenerator{new(unitStrGener), nil, nil},
+			geners:        []dataGenerator{&unitStrGener{}, nil, nil},
 		},
 	},
 	ast.TimestampDiff:    {},