@shihongzhi 2019-10-30
> It seems that there is some thing wrong in your code. Please fix the CI fail.

```
[2019-10-30T12:37:11.441Z] FAIL: builtin_time_vec_test.go:153: testEvaluatorSuite.TestVectorizedBuiltinTimeFunc
[2019-10-30T12:37:11.441Z] 
[2019-10-30T12:37:11.441Z] builtin_time_vec_test.go:154:
[2019-10-30T12:37:11.441Z]     testVectorizedBuiltinFunc(c, vecBuiltinTimeCases)
[2019-10-30T12:37:11.441Z] bench_test.go:926:
[2019-10-30T12:37:11.441Z]     c.Assert(2*vecWarnCnt, Equals, totalWarns)
[2019-10-30T12:37:11.441Z] ... obtained uint16 = 0x0
[2019-10-30T12:37:11.441Z] ... expected uint16 = 0x151
```
I don't know why have this problem. Could you help me?